requirements.txt
langchain
langchain_community
unstructured
libmagic
python-magic
python-magic-bin
faiss-cpu
sentence-transformers
pandas
numpy

streamlit
os
pickle5
time
python-dotenv
openai
tiktoken